+++ test/testUnitConverter.cpp
@@ -44,8 +44,9 @@
  */
 template < Quantity _quantity_ >
 void check_quantity(std::vector< std::string > unitnames,
-                    std::vector< int > unitlogic) {
-  for (unsigned int i = 0; i < unitnames.size(); ++i) {
+                    std::vector< int_fast32_t > unitlogic) {
+
+  for (size_t i = 0; i < unitnames.size(); ++i) {
     assert_condition(UnitConverter::convert(1., unitnames[i], unitnames[i]) ==
                      1.);
     assert_condition(UnitConverter::to_unit< _quantity_ >(
@@ -83,6 +84,7 @@ void check_quantity(std::vector< std::string > unitnames,
  * @return Exit code: 0 on success.
  */
 int main(int argc, char **argv) {
+
   {
     Unit unit = UnitConverter::get_unit(" m");
     Unit unit2(1., 1, 0, 0, 0, 0, 0);
@@ -111,7 +113,7 @@ int main(int argc, char **argv) {
   /// ACCELERATION
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m s^-2");
     unitlogic.push_back(0);
 
@@ -121,7 +123,7 @@ int main(int argc, char **argv) {
   /// ANGLE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("radians");
     unitlogic.push_back(0);
     unitnames.push_back("degrees");
@@ -134,7 +136,7 @@ int main(int argc, char **argv) {
   /// DENSITY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("kg m^-3");
     unitlogic.push_back(0);
     unitnames.push_back("g cm^-3");
@@ -147,7 +149,7 @@ int main(int argc, char **argv) {
   /// ENERGY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("J");
     unitlogic.push_back(0);
     unitnames.push_back("eV");
@@ -161,7 +163,7 @@ int main(int argc, char **argv) {
   /// ENERGY CHANGE RATE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("kg m^-1s^-3");
     unitlogic.push_back(0);
     unitnames.push_back("J m^-3s^-1");
@@ -178,7 +180,7 @@ int main(int argc, char **argv) {
   /// ENERGY_RATE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("kg m^2s^-3");
     unitlogic.push_back(0);
     unitnames.push_back("J s^-1");
@@ -194,7 +196,7 @@ int main(int argc, char **argv) {
   /// FLUX
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m^-2s^-1");
     unitlogic.push_back(0);
     unitnames.push_back("cm^-2s^-1");
@@ -208,7 +210,7 @@ int main(int argc, char **argv) {
   /// FREQUENCY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("s^-1");
     unitlogic.push_back(0);
     unitnames.push_back("Hz");
@@ -224,7 +226,7 @@ int main(int argc, char **argv) {
   /// FREQUENCY PER MASS
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("s^-1 kg^-1");
     unitlogic.push_back(0);
     unitnames.push_back("s^-1 Msol^-1");
@@ -238,7 +240,7 @@ int main(int argc, char **argv) {
   /// LENGTH
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m");
     unitlogic.push_back(0);
     unitnames.push_back("cm");
@@ -260,7 +262,7 @@ int main(int argc, char **argv) {
   /// MASS
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("kg");
     unitlogic.push_back(0);
     unitnames.push_back("g");
@@ -276,7 +278,7 @@ int main(int argc, char **argv) {
   /// MASS RATE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("kg s^-1");
     unitlogic.push_back(0);
     unitnames.push_back("Msol yr^-1");
@@ -290,7 +292,7 @@ int main(int argc, char **argv) {
   /// NUMBER DENSITY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m^-3");
     unitlogic.push_back(0);
     unitnames.push_back("cm^-3");
@@ -304,7 +306,7 @@ int main(int argc, char **argv) {
   /// OPACITY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m^-1");
     unitlogic.push_back(0);
 
@@ -314,7 +316,7 @@ int main(int argc, char **argv) {
   /// REACTION RATE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m^3s^-1");
     unitlogic.push_back(0);
     unitnames.push_back("cm^3s^-1");
@@ -327,7 +329,7 @@ int main(int argc, char **argv) {
   /// SURFACE AREA
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m^2");
     unitlogic.push_back(0);
     unitnames.push_back("cm^2");
@@ -340,7 +342,7 @@ int main(int argc, char **argv) {
   /// TEMPERATURE
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("K");
     unitlogic.push_back(0);
 
@@ -350,7 +352,7 @@ int main(int argc, char **argv) {
   /// TIME
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("s");
     unitlogic.push_back(0);
     unitnames.push_back("Gyr");
@@ -366,7 +368,7 @@ int main(int argc, char **argv) {
   /// VELOCITY
   {
     std::vector< std::string > unitnames;
-    std::vector< int > unitlogic;
+    std::vector< int_fast32_t > unitlogic;
     unitnames.push_back("m s^-1");
     unitlogic.push_back(0);
     unitnames.push_back("km s^-1");
